feat(contextselector): splitting comp to provider an element
