+++ storybook/src/stories/context-selector/component.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, Story, Canvas, ArgsTable, Description } from '@storybook/addon-docs';
 import ChangeLog from '../../components/ChangeLog';
 
-import { ContextSelector, ContextSelectorProps } from '@equinor/fusion-react-context-selector';
+import { ContextProvider, ContextSelector, ContextSelectorProps } from '@equinor/fusion-react-context-selector';
 import { _handleSelect, _exampleResolver } from '../searchable-dropdown/component.helpers.tsx';
 
 <Meta title="Data/ContextSelector" />
@@ -15,7 +15,9 @@ import { _handleSelect, _exampleResolver } from '../searchable-dropdown/componen
 export const Template = ({ children, ...props }) => {
   return (
     <div style={{ width: '98%', maxWidth: '420px', height: '350px', margin: '0 auto' }}>
-      <ContextSelector {...props}>{children}</ContextSelector>
+      <ContextProvider resolver={_exampleResolver}>
+        <ContextSelector {...props}>{children}</ContextSelector>
+      </ContextProvider>
     </div>
   );
 };
@@ -23,13 +25,11 @@ export const Template = ({ children, ...props }) => {
 ## ContextSelector Component 
 <Canvas withSource="closed">
   <Story name="Component" args={{
-      label: 'Context selector',
       placeholder: 'Start to type to search...',
       initialText: 'The initial text result',
       id: 'sdd-test',
       variant: 'page',
       dropdownHeight: '300px',
-      resolver: _exampleResolver
     }}
     argTypes={{
       id: {
@@ -55,6 +55,7 @@ export const Template = ({ children, ...props }) => {
         description: 'The scroll height for dropdown result list'
       },
       label: {
+        type: 'string',
         description: 'Textinput label'
       },
       value: {
@@ -76,6 +77,79 @@ export const Template = ({ children, ...props }) => {
 </Canvas>
 
 <ArgsTable story="Component" />
+
+
+## How to use
+
+The ``ContextSelector`` component is built up of 2 elements, ``ContextProvider`` and ``ContextSelector``.
+
+Wrap the Selector(s) in the Provider and add a resolver to the provider element by using the attribute resolver.
+The resolver must be of type ``ContextResolver``.
+
+**Example usage**
+
+```ts
+const myResolver: ContextResolver = {
+  searchQuery: async (query: string) => {
+    // API call
+  },
+  initialResult: [
+    // pre search available ContextResult
+  ],
+  closeHandler: (e: MouseEvent) => {
+    // callback after UI close dropdown list
+  },
+};
+const Context = (props): JSX.Element => {
+  return (
+    <ContextProvider resolver={myResolver}>
+      // other element
+      <ContextSelector {...props} />
+    </DropdownProvider>
+  );
+};
+```
+
+## Resolver methods
+
+**searchQuery**
+
+The resolver method ``searchQuery`` is called on every keyup event with 250ms debounce from the textinput element. So place your query call on this method.
+
+Your method should return a Promise which resolves to type ``ContextResult`` which is an array of ``ContextResultItem`` with the following properties.
+
+```typescript
+interface ContextResultItem {
+  id: string;
+  title?: string;
+  subTitle?: string;
+  isDisabled?: boolean;
+  isSelected?: boolean;
+  isError?: boolean;
+  meta?: string;
+  graphic?: string;
+  type?: 'section' | 'divider' | null;
+  children?: ContextResult;
+}
+```
+
+If you want sectioned items, use ``type: 'section`` and adds its children to the ``children`` property.
+
+See storybook story for examples.
+
+**initialResult**
+
+You can pass ``ContextResult`` to ``initialResult`` to show in the list before any search is performed. Use this for Favourites or most used context's.
+
+**closeHandler**
+
+The web-component has a handler for closing the list and clearing value in textinput field, but you can pass an additional callback to this handler. Use this for clearing context in your app.
+
+## onSelect
+
+When the user selects an item from the result list the ``onSelect`` event gets triggered. The ``onSelect`` events detail object contains a ``ContextResult`` of the selected item, or items if multiple flag is set.
+
+The event is of type ``ContextSelectEvent``.
 
 ## Extending component
