+++ storybook/src/stories/searchable-dropdown/component.stories.mdx
@@ -4,7 +4,7 @@ import { Dropdown, DropdownProvider, useDropdownProviderRef } from '@equinor/fus
 import { _handleAction, _exampleResolver } from './component.helpers';
 
 <Meta
-  title="Data/Searchable-Dropdown"
+  title="Data/SearchableDropdown"
   component={Dropdown}
 />
